refactor(policy-types): collapse 3-way duplicated prepaid policies
Extract PrepaidBalancePerRequestPolicy and PrepaidBalancePerDocumentPolicy
generic bases under policy_types/prepaid/. Concrete provider policies
(Stripe, Xendit) become 5-line declarations of NAME, DESCRIPTION,
PROVIDER_NAME, and CONFIG_CLS. All hook bodies, config validation,
schema export, and tier-matching logic now live once in the shared
package.

Also collapse the per-provider PaymentConfig classes (StripePerRequestConfig,
XenditPerRequestConfig, etc.) into shared PrepaidPerRequestConfig and
PrepaidPerDocumentConfig in prepaid/policy_config.py. They were byte-
identical apart from class names; the Stripe variant's own docstring
called out that it "mirrors Xendit's behavior."

Net: ~600 LOC of duplication absorbed. The per-provider classes that
remain (StripePerRequestPolicy, etc.) are the irreducible declarative
kernel of the registry pattern — distinct NAME, distinct
required_wallet_type, distinct display string — not duplication.

MPP policies are intentionally unchanged: MppCharger has a different
Protocol (Challenge/Credential, not reserve/cancel), so they can't share
the prepaid hook bodies.

Adding a 4th prepaid gateway is now two ~6-line subclass files plus a
PREPAID_BALANCE_WALLET_TYPES entry — no copy-paste of hook bodies.
