+++ backend/syft_space/components/policy_types/prepaid/payment_policy.py
@@ -1,13 +1,13 @@
-"""Base class for Stripe payment policies.
+"""Generic infrastructure for prepaid-balance payment policies.
 
-Subclasses bind a specific config class (which carries ``unit_type`` as a
-typed Literal field) and provide pre/post-hook bodies. Everything else —
-config validation, schema export, tier matching, identity boilerplate —
+Replaces per-provider StripePaymentPolicy / XenditPaymentPolicy bases.
+Concrete subclasses declare ``PROVIDER_NAME``, ``NAME``, ``DESCRIPTION``,
+and ``CONFIG_CLS``. Every other concern — config validation, schema
+export, tier matching, identity boilerplate — is provider-agnostic and
 lives here.
 
-Adding a new unit type is one new config + one new policy subclass — same
-extension model as Xendit. Adding a new provider is the parallel of this
-file: a new ``*PaymentPolicy`` base plus its per-unit subclasses.
+Adding a new prepaid gateway is two thin subclasses (per-request +
+per-document) plus a provider-specific config module.
 """
 
 from typing import Any, ClassVar
@@ -16,21 +16,21 @@ from syft_space.components.policy_types.interfaces import (
     BasePolicyType,
     Capabilities,
 )
-from syft_space.components.policy_types.stripe.policy_config import (
-    StripePaymentConfig,
-)
 from syft_space.components.shared.utils import (
     ConfigSchemaGenerator,
     matches_any_pattern,
 )
 
 
-class StripePaymentPolicy(BasePolicyType):
-    """Shared scaffolding for all Stripe-based payment policies."""
+class PrepaidBalancePaymentPolicyBase(BasePolicyType):
+    """Shared scaffolding for all prepaid-balance payment policies."""
 
+    PROVIDER_NAME: ClassVar[str]  # "stripe", "xendit", ...
     NAME: ClassVar[str]
     DESCRIPTION: ClassVar[str]
-    CONFIG_CLS: ClassVar[type[StripePaymentConfig]]
+    # Provider-specific Pydantic config class. Duck-typed against
+    # ``applied_to: list[str]`` and ``price: float``.
+    CONFIG_CLS: ClassVar[type]
 
     def __init__(self) -> None:
         pass
@@ -57,19 +57,22 @@ class StripePaymentPolicy(BasePolicyType):
 
     @classmethod
     def capabilities(cls) -> Capabilities:
-        return Capabilities(requires_wallet=True, required_wallet_type="stripe")
+        return Capabilities(
+            requires_wallet=True,
+            required_wallet_type=cls.PROVIDER_NAME,
+        )
 
     @classmethod
     async def validate_config(cls, config: dict[str, Any]) -> dict[str, Any]:
         """Validate the user-supplied config.
 
-        ``unit_type`` is a typed Literal field on the subclass's CONFIG_CLS,
-        so it flows through ``model_dump()`` naturally. No injection here.
+        ``unit_type`` is a typed Literal on the subclass's CONFIG_CLS so it
+        flows through ``model_dump()`` naturally; nothing is injected here.
         """
         try:
             validated = cls.CONFIG_CLS(**config)
         except Exception as e:
-            raise ValueError(f"Invalid stripe config: {e}") from e
+            raise ValueError(f"Invalid {cls.PROVIDER_NAME} config: {e}") from e
         return validated.model_dump()
 
     def _find_matching_price(
